remove asynchronous web calls, vastly improving reliability

+++ src/main/java/com/northpine/scrape/ogr/OgrCollector.java
@@ -2,13 +2,18 @@ package com.northpine.scrape.ogr;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.slf4j.Marker;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStreamReader;
+import java.nio.channels.Channels;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Scanner;
 import java.util.concurrent.CompletableFuture;
+import java.util.stream.Collectors;
 
 public abstract class OgrCollector {
 
@@ -26,25 +31,32 @@ public abstract class OgrCollector {
     this.extension = extension;
   }
 
-  public synchronized void addJsonToPool(String file) {
-    ProcessBuilder builder = new ProcessBuilder( "ogr2ogr", "-f", ""+ ogrFormat + "", "-append", poolBase + extension, file );
+  public synchronized boolean addJsonToPool(String file) {
+    ProcessBuilder builder = new ProcessBuilder( "ogr2ogr", "-f", ogrFormat, "-append", poolBase + extension, file );
     try {
       Process p;
       p = builder.start();
-      p.waitFor();
-      Scanner c = new Scanner(p.getErrorStream());
-      while(c.hasNext()) {
-        log.error(c.nextLine());
+      int status = p.waitFor();
+      var error = new BufferedReader(new InputStreamReader(p.getErrorStream()))
+          .lines().parallel()
+          .collect(Collectors.joining("\n")
+      );
+      if(error.length() > 0) log.warn(error);
+      if(status == 0) {
+        CompletableFuture.runAsync( () -> {
+          try {
+            Files.delete( Paths.get( file ) );
+          } catch ( IOException e ) {
+            log.warn("Couldn't delete " + file, e);
+          }
+        } );
+        return true;
+      } else {
+        return false;
       }
-      CompletableFuture.runAsync( () -> {
-        try {
-          Files.delete( Paths.get( file ) );
-        } catch ( IOException e ) {
-          log.warn("Couldn't delete " + file, e);
-        }
-      } );
     } catch ( IOException | InterruptedException e ) {
       log.error("ogr2ogr failed", e);
+      return false;
     }
   }
